backend/requirement.txt
fastapi
uvicorn
sqlmodel
pymysql
requests
python-dotenv
tiktoken